+++ src/zenml/artifact_stores/base_artifact_store.py
@@ -35,6 +35,7 @@ from zenml.enums import StackComponentType
 from zenml.exceptions import ArtifactStoreInterfaceError
 from zenml.io import fileio
 from zenml.stack import Flavor, StackComponent, StackComponentConfig
+from zenml.utils import io_utils
 
 PathType = Union[bytes, str]
 
@@ -60,7 +61,7 @@ def _sanitize_potential_path(potential_path: Any) -> Any:
         # Neither string nor bytes, this is not a path
         return potential_path
 
-    if fileio.is_remote(path):
+    if io_utils.is_remote(path):
         # If we have a remote path, replace windows path separators with
         # slashes
         import ntpath
